efi: cper: Update Zhaoxin/Centaur ZDI/ZPI error decode
zhaoxin inclusion
category: feature
bugzilla: https://atomgit.com/openeuler/kernel/issues/8948
CVE: NA

(cherry-pick from OLK-6.6 commit 1ea20cb16d1f)

-------------------

Add support for decoding ZDI/ZPI error information on Zhaoxin/Centaur CPUs:

1. ZPI (socket interconnection) and ZDI (die interconnection) error
   handling for Zhaoxin/Centaur processors

2. Enhanced error reporting with human-readable error messages for
   ZDI/ZPI errors through CPER (Common Platform Error Record)

3. Fix function parameter type for cper_zdi_zpi_err_type_str()
   from unsigned long to unsigned int

4. Add proper X86 platform condition checks for ZDI/ZPI error handling

When ZDI or ZPI errors occur, they trigger SMI interrupts which fill
error information into CPER structures. This change improves the
readability of these error messages for Zhaoxin/Centaur CPUs.

+++ include/linux/cper.h
@@ -598,6 +598,6 @@ void cper_estatus_print(const char *pfx,
 int cper_estatus_check_header(const struct acpi_hest_generic_status *estatus);
 int cper_estatus_check(const struct acpi_hest_generic_status *estatus);
 
-const char *cper_zdi_zpi_err_type_str(unsigned long etype);
+const char *cper_zdi_zpi_err_type_str(unsigned int etype);
 
 #endif
